requirements.txt
flask==2.3.3
flask-cors==4.0.0
pymongo==4.5.0
python-telegram-bot==20.6
aiogram==3.1.1
python-dotenv==1.0.0
requests==2.31.0
cryptography==41.0.7
pyjwt==2.8.0
bcrypt==4.0.1
asyncio==3.4.3
aiohttp==3.8.6
motor==3.3.2
dnspython==2.4.2
